Reduced dithering and rastering in Amstrad/Commodore filters Removed Otsu thresholding since it caused chronic bad exposure Keep multiple copies of images processed from gallery

+++ src/se/embargo/retroboy/color/IColorDistance.java
@@ -4,5 +4,8 @@ package se.embargo.retroboy.color;
  * Returns the distance between two colors.
  */
 public interface IColorDistance {
-	public int get(int r1,int g1,int b1, int r2,int g2,int b2);
+	/**
+	 * @return	A number between 0 and 1 indicating the color distance.
+	 */
+	public double get(int r1,int g1,int b1, int r2,int g2,int b2);
 }
